Use 64-bit integer for channel IDs
The javascript version already had this. Since we're redefining almost everywhere that channel IDs are used as part of the v3 of the protocol, it seems a reasonable time to upgrade to 64-bit integers on the .NET side as well.

+++ src/Nerdbank.Streams/MultiplexingStream.FrameHeader.cs
@@ -23,7 +23,7 @@ namespace Nerdbank.Streams
             /// <summary>
             /// Gets or sets the channel that this frame refers to or carries a payload for.
             /// </summary>
-            internal int? ChannelId { get; set; }
+            internal long? ChannelId { get; set; }
 
             /// <summary>
             /// Gets or sets a value indicating whether <see cref="ChannelId"/> is referring to a channel that was originally offered by the sender of this frame.
@@ -37,7 +37,7 @@ namespace Nerdbank.Streams
             /// Gets the ID of the channel that this frame refers to or carries a payload for.
             /// </summary>
             /// <exception cref="MultiplexingProtocolException">Thrown if <see cref="ChannelId"/> is null.</exception>
-            internal int RequiredChannelId => this.ChannelId ?? throw new MultiplexingProtocolException("Expected ChannelId not present in frame header.");
+            internal long RequiredChannelId => this.ChannelId ?? throw new MultiplexingProtocolException("Expected ChannelId not present in frame header.");
 
             /// <summary>
             /// Gets the text to display in the debugger when an instance of this struct is displayed.
